requirements.txt
beautifulsoup4
Jinja2
markdown-it-py[linkify,plugins]
mypy
pandas
pygments
pyjanitor
requests
types-requests
